requirements.txt
dj-database-url==0.5.0
Django==2.2.24
django-environ==0.7.0
django-filter==2.4.0
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.3.0
